Refactor stuff to `common.py`
